Separated constants into different files
