--- src/main/java/frc/robot/Constants.java
@@ -1,18 +1,0 @@
-package frc.robot;
-
-public class Constants {
-    public class DrivetrainConstants {
-        public static final int kLeftFrontPort=0;
-        public static final int kLeftBackPort=1;
-        public static final int kLeftTopPort=2;
-
-        public static final int kRightFrontPort=3;
-        public static final int kRightBackPort=4;
-        public static final int kRightTopPort=3;
-    }
-
-    public class XBConstants {
-        public static final int drivePort = 0;
-        public static final int opPort = 1;
-    }
-}
