+++ src/main/java/frc/robot/constants/DrivetrainConstants.java
@@ -1,0 +1,11 @@
+package frc.robot.constants;
+
+public class DrivetrainConstants {
+    public static final int kLeftFrontPort=0;
+    public static final int kLeftBackPort=1;
+    public static final int kLeftTopPort=2;
+
+    public static final int kRightFrontPort=3;
+    public static final int kRightBackPort=4;
+    public static final int kRightTopPort=3;  
+}
